@@ -8,49 +8,85 @@
     "build": "ionic-app-scripts build",
     "clean": "ionic-app-scripts clean",
     "ionic:build": "ionic-app-scripts build",
-    "ionic:serve": "ionic-app-scripts serve"
+    "ionic:serve": "ionic-app-scripts serve",
+    "setup_powershell": "cd src; git clone ",
+    "postsetup_powershell": "cd src; rename savi-lib lib; cd lib; (git remote add upstream https://github.com/ml5174/savi-lib.git || echo skipping, moving on); git checkout master; git pull upstream master; cd ../..; (git remote add upstream https://github.com/ml5174/sacpcp.git || echo skipping, moving on); git checkout master; git pull upstream master; npm install -g ionic cordova; npm install; ionic serve",
+    "setup": "cd src && git clone ",
+    "postsetup": "cd src && (rename savi-lib lib || mv savi-lib lib) && cd lib && (git remote add upstream https://github.com/ml5174/savi-lib.git || echo skipping, moving on) && git checkout master && git pull upstream master && cd ../.. && (git remote add upstream https://github.com/ml5174/sacpcp.git || echo skipping, moving on) && git checkout master && git pull upstream master && npm install -g ionic cordova && npm install && ionic serve"
   },
   "dependencies": {
-    "@angular/common": "2.2.1",
-    "@angular/compiler": "2.2.1",
-    "@angular/compiler-cli": "2.2.1",
-    "@angular/core": "2.2.1",
-    "@angular/forms": "2.2.1",
-    "@angular/http": "2.2.1",
-    "@angular/platform-browser": "2.2.1",
-    "@angular/platform-browser-dynamic": "2.2.1",
-    "@angular/platform-server": "2.2.1",
-    "@ionic/storage": "1.1.7",
-    "ionic-angular": "2.0.0-rc.5-201701112208",
-    "ionic-native": "2.2.11",
+    "@angular/animations": "5.2.7",
+    "@angular/common": "5.2.7",
+    "@angular/compiler": "5.2.7",
+    "@angular/compiler-cli": "5.2.7",
+    "@angular/core": "5.2.7",
+    "@angular/forms": "5.2.7",
+    "@angular/http": "5.2.7",
+    "@angular/platform-browser": "5.2.7",
+    "@angular/platform-browser-dynamic": "5.2.7",
+    "@ionic-native/app-version": "^4.5.3",
+    "@ionic-native/core": "4.5.3",
+    "@ionic-native/dialogs": "4.5.3",
+    "@ionic-native/in-app-browser": "^4.9.1",
+    "@ionic-native/splash-screen": "^4.5.3",
+    "@ionic-native/status-bar": "^4.5.3",
+    "@ionic/storage": "2.1.3",
+    "@ngx-translate/core": "9.1.1",
+    "@ngx-translate/http-loader": "^2.0.1",
+    "angular2-text-mask": "8.0.4",
+    "cordova-android": "6.3.0",
+    "cordova-ios": "4.5.5",
+    "cordova-plugin-app-version": "^0.1.9",
+    "cordova-plugin-appversion": "^1.0.0",
+    "cordova-plugin-device": "^1.1.7",
+    "cordova-plugin-dialogs": "^2.0.1",
+    "cordova-plugin-hide-home-indicator": "^1.0.1",
+    "cordova-plugin-inappbrowser": "^3.0.0",
+    "cordova-plugin-ionic-webview": "^1.2.1",
+    "cordova-plugin-splashscreen": "^4.1.0",
+    "cordova-plugin-statusbar": "2.4.1",
+    "cordova-plugin-whitelist": "^1.3.3",
+    "ionic-angular": "^3.9.2",
+    "ionic-plugin-deeplinks": "^1.0.17",
+    "ionic-plugin-deploy": "^0.6.7",
+    "ionic-plugin-keyboard": "^2.2.1",
     "ionicons": "3.0.0",
-    "moment": "^2.17.0",
-    "moment-timezone": "^0.5.10",
-    "ng2-translate": "^3.0.0",
-    "rxjs": "5.0.0-beta.12",
-    "sw-toolbox": "3.4.0",
-    "zone.js": "0.6.26"
+    "moment": "2.20.1",
+    "moment-timezone": "0.5.14",
+    "rxjs": "^5.5.6",
+    "sw-toolbox": "3.6.0",
+    "zone.js": "0.8.18"
   },
   "devDependencies": {
-    "@ionic/app-scripts": "1.0.0",
-    "@types/moment-timezone": "^0.2.32",
-    "typescript": "2.0.9"
+    "@angular/tsc-wrapped": "^4.4.6",
+    "@ionic/app-scripts": "^3.1.8",
+    "@types/moment-timezone": "^0.5.4",
+    "typescript": "~2.4.2"
   },
-  "cordovaPlugins": [
-    "cordova-plugin-device",
-    "cordova-plugin-console",
-    "cordova-plugin-whitelist",
-    "cordova-plugin-splashscreen",
-    "cordova-plugin-statusbar",
-    "ionic-plugin-keyboard"
-  ],
-  "cordovaPlatforms": [
-    "ios",
-    {
-      "platform": "ios",
-      "version": "",
-      "locator": "ios"
-    }
-  ],
-  "description": "savi: An Ionic project"
+  "description": "savi: An Ionic project",
+  "cordova": {
+    "plugins": {
+      "cordova-plugin-app-version": {},
+      "cordova-plugin-device": {},
+      "cordova-plugin-splashscreen": {},
+      "cordova-plugin-statusbar": {},
+      "cordova-plugin-whitelist": {},
+      "ionic-plugin-deploy": {},
+      "ionic-plugin-keyboard": {},
+      "ionic-plugin-deeplinks": {
+        "URL_SCHEME": "tsadfw",
+        "DEEPLINK_SCHEME": "https",
+        "DEEPLINK_HOST": "https://www.volunteers.tsadfw.org",
+        "ANDROID_PATH_PREFIX": "/"
+      },
+      "cordova-plugin-ionic-webview": {},
+      "cordova-plugin-dialogs": {},
+      "cordova-plugin-hide-home-indicator": {},
+      "cordova-plugin-inappbrowser": {}
+    },
+    "platforms": [
+      "android",
+      "ios"
+    ]
+  }
 }
