@@ -3,8 +3,8 @@ import { UserServices } from '../../lib/service/user';
 import { NavController, NavParams, PopoverController } from 'ionic-angular';
 import { STRINGS } from '../../lib/provider/config';
 import { ConfirmEmailPage } from '../confirm-email/confirm-email';
-import { TranslateService } from "ng2-translate/ng2-translate";
-import { HomePage } from '../home/home';
+import { TranslateService } from "@ngx-translate/core";
+import { LoginPage } from '../login/login';
 import { PasswordPopover } from '../../popover/password';
 
 @Component({
@@ -17,18 +17,18 @@ export class ChangePasswordPage {
   public password2: string = '';
   public email: string = '';
   public sms: string = '';
-
   public passworderror: boolean = false;
   public password1error: boolean = false;
   public password2error: boolean = false;
   public emailerror: boolean = false;
   public smserror: boolean = false;
-
   public key: string = '';
   public val: string = '';
   public errors: Array<string> = [];
   public pcmethod: string = 'email'
   public pcvalue: string = '';
+  public meetsRequirement = false;
+  public MINPWLENGTH = 1;
 
   constructor(public nav: NavController,
     public navParams: NavParams,
@@ -38,8 +38,15 @@ export class ChangePasswordPage {
       console.dir(navParams);
 
   }
+
+  checkRequirement($event) {
+    if (this.password1.length >= this.MINPWLENGTH && this.password2.length >= this.MINPWLENGTH) {
+      this.meetsRequirement = true;
+      return;
+    }
+    this.meetsRequirement = false;
+  }
   register() {
-    let registerLogin = this;
     this.errors = [];
     this.passworderror = false;
     this.password1error = false;
@@ -100,7 +107,7 @@ export class ChangePasswordPage {
 
   }
   back() {
-    this.nav.pop(HomePage);
+    this.nav.push(LoginPage);
   }
   showPassword() {
     if (this.showpassword === 'password') this.showpassword = 'text';
