@@ -1,17 +1,17 @@
-import { Component , ViewChild} from '@angular/core'
+import { Component, ViewChild } from '@angular/core'
 import { UserServices } from '../../lib/service/user';
-import { NavController, NavParams, PopoverController } from 'ionic-angular';
+import { NavController, PopoverController, ModalController, Events, NavParams } from 'ionic-angular';
 import { STRINGS } from '../../lib/provider/config';
-import { TermsPage } from '../terms/terms';
-import { ConfirmEmailPage } from '../confirm-email/confirm-email';
-import { ConfirmSMSPage } from '../confirm-sms/confirm-sms';
-import { TranslateService } from "ng2-translate/ng2-translate";
+import { TranslateService } from "@ngx-translate/core";
 import { HomePage } from '../home/home';
 import { RegisterIndividualProfilePage } from '../register-individual-profile/register-individual-profile';
 import { PasswordPopover } from '../../popover/password';
 import { UseridPopover } from '../../popover/userid';
+import { PrivacyTermsModal } from '../../modals/privacy-terms-modal';
 import { Storage } from '@ionic/storage';
 import { ContactMethod } from '../../lib/components/ContactMethod/contactMethod.component';
+import { Content } from 'ionic-angular';
+
 @Component({
   templateUrl: 'register-login.html'
 })
@@ -28,32 +28,199 @@ export class RegisterLoginPage {
   public password2error: boolean = false;
   public emailerror: boolean = false;
   public smserror: boolean = false;
+  public termserror: boolean = false;
+  public nonFieldsError: boolean = false;
 
+  // Error values
+  public usernameerrorvalue: string = 'This field is required.';
+  public password1errorvalue: string = 'This field is required.';
+  public password2errorvalue: string = 'This field is required.';
+  public emailerrorvalue: string = '';
+  public smserrorvalue: string = '';
+  public termserrorvalue: string = "You must accept Privacy and Terms to proceed.";
+  public nonFieldsErrorValue = '';
+
+  private modalClicked: boolean = false;
   public key: string = '';
   public val: string = '';
   public errors: Array<string> = [];
-  /* Move the email, Phone contact details to a common component -
-     ViewChild(ContactMethod)
-     This can be reused at multiple pages.
-   */
-  //public pcmethod: string = 'email'
-  //public pcvalue: string = '';
-  //public mobileNumberAreaCode = '';
-  //public mobileNumberPrefix = '';
- // public mobileNumberLineNumber = '';
+
+  @ViewChild(Content) content: Content;
   @ViewChild(ContactMethod)
   public contactMethod: ContactMethod;
-
   public terms: boolean = false;
   public remember: boolean = true;
-  public storage: Storage = new Storage();
   public pcmethod: string = 'email'
-
+  public meetsRequirement: boolean = false;
   constructor(public nav: NavController,
+    public storage: Storage,
     public userServices: UserServices,
     public translate: TranslateService,
-    public popoverCtrl: PopoverController
-  ) { }
+    public popoverCtrl: PopoverController,
+    public modalCtrl: ModalController,
+    private navParams: NavParams,
+    private navCtrl:NavController,
+    private ev: Events
+  ){}
+
+  promiseToScroll() {
+    //needed to allow view to refresh with error elements before scroll
+    console.log("returning promise to life");
+    return new Promise((resolve, reject) => {
+      setTimeout(() => {
+        resolve(42);
+      }, 200);
+    });
+  }
+
+  checkContactMethod() {
+    console.log("checking contact method");
+      this.emailerror=false;
+      this.smserror=false;
+    if(this.contactMethod.pcmethod==='email') {
+      let email = this.contactMethod.pcvalue;
+      if(email) {
+        let at=email.indexOf('@');
+        let dot=email.lastIndexOf('.');
+        let length=email.length;
+        if(at===-1 || dot===-1 || dot<at || length-1===dot) {
+          this.emailerror=true;
+          this.emailerrorvalue='Enter a valid email address.'
+        }
+      } else {
+        this.emailerror=true;
+        this.emailerrorvalue='This field is may not be empty.';
+      }
+    } else {
+      if(!this.contactMethod.mobilenumber) {
+        this.smserror=true;
+        this.smserrorvalue='This field may not be empty.';
+      }
+      if(this.contactMethod.mobilenumber.getPN().length<11) {
+        this.smserror=true;
+        this.smserrorvalue='Mobile number should be 10 digits.';
+      }
+    }
+    return this.emailerror || this.smserror;
+  }
+
+  checkPasswordsMatching() {
+      console.log("checkPasswordsMatching()");
+    this.password2error=false;
+    this.password2errorvalue = '';
+    if(!this.password2 || this.password2.trim().length == 0) {
+        console.log("Oh noes!!: " + this.password2);
+        this.password2error = true;
+        this.password2errorvalue = 'This field may not be empty.';
+    }
+    else if(this.password1 && this.password1!=='') {
+      if(this.password1!==this.password2) {
+        this.meetsRequirement=false;
+        this.password2error=true;
+        this.password2errorvalue='Your passwords do not match.';
+      }
+    }
+    return this.password2error;
+  }
+
+  clearContactErrors() {
+    this.emailerror=false;
+    this.smserror=false;
+  }
+
+  checkUsername() {
+    this.usernameerror=false;
+    var beginsWithLetterConstraint = new RegExp('^[A-Za-z]');
+    var beginsWithLetter = this.username.match(beginsWithLetterConstraint);
+    var badUsernameCharacters = new RegExp('[^\+\-\.@A-Z_a-z0-9]');
+    var hasBadCharactersInUsername = this.username.match(badUsernameCharacters);
+    if(!this.username || this.username==='') {
+      this.usernameerror=true;
+      this.usernameerrorvalue='This field may not be empty.';
+    }else if(this.username.length<8) {
+      this.usernameerror=true;
+      this.usernameerrorvalue='Username must be at least 8 characters.';
+    }else if(this.username.length>30) {
+      this.usernameerror=true;
+      this.usernameerrorvalue='Username must be less than 30 characters';
+    }else if(!beginsWithLetter) {
+      this.usernameerror=true;
+      this.usernameerrorvalue='Username must begin with a letter';
+    }
+    else if(hasBadCharactersInUsername) {
+      this.usernameerror=true;
+      this.usernameerrorvalue='Username must only contain alphanumeric characters, @, +, ., _, and -';
+    }
+    return this.usernameerror;
+  }
+
+  checkPassword1() {
+    this.password1error=false;
+    if(!this.password1) {
+      this.password1error=true;
+      this.password1errorvalue='This field may not be empty.';
+    } else if(this.password1.length<8) {
+      this.password1error=true;
+      this.password1errorvalue='Password must be at least 8 characters.';
+    } else if(this.password1===this.username) {
+      this.password1error=true;
+      this.password1errorvalue='Password must not match username.';
+    } else if(!this.complexPassword()) {
+      this.password1error=true;
+      this.password1errorvalue= 'This password does not meet complexity requirements.';//'Password must contain at least 3 of the following characteristics: 1. Uppercase English letters\n2. Lowercase English letters\n3. Numbers\n4. Non-alphanumeric characters \" # $ % & \' ( ) * + , - . : ; < = > ? @ [ \\ ] ^ { }';
+    }
+    return this.password1error;
+  }
+
+  complexPassword() {
+    var uppercase = new RegExp('[A-Z]');
+    var lowercase = new RegExp('[a-z]');
+    var numbers = new RegExp('[0-9]');
+    var specialCharacters = new RegExp('[" # $ % & \' ( ) * + , - . : ; < = > ? @ [ \\ ] ^ { }]');
+    var numMatches = 0;
+    if(this.password1.match(uppercase))
+      numMatches++;
+    if(this.password1.match(lowercase))
+      numMatches++;
+    if(this.password1.match(numbers))
+      numMatches++;
+    if(this.password1.match(specialCharacters))
+      numMatches++;
+    return (numMatches >= 3);
+  }
+
+  checkRequired() {
+    // sets the variable meetsRequirement to true if required fields are not empty
+    // contactMethod prefix 1 to the numbers entered
+    //console.log("check rquired called, preffered contact method is ", this.contactMethod.pcmethod);
+    if (this.contactMethod.pcmethod === 'email') {
+      if (this.username && this.password1 && this.password2 && this.contactMethod.pcvalue && this.terms) {
+        this.meetsRequirement = true;
+        return true;
+      }
+      this.meetsRequirement = false;
+      return false;
+    }
+    if (this.contactMethod.mobilenumber) {
+      if (this.username && this.password1 && this.password2 && (this.contactMethod.mobilenumber.getPN().length >= 11) && this.terms) {
+        this.meetsRequirement = true;
+        //console.log("inside contactMEthodMobileNumber not falsey", this.contactMethod.mobilenumber.getPN());
+        return true;
+      }
+    }
+    this.meetsRequirement = false;
+    return false;
+  }
+
+  checkAll() {
+    let pw1 = this.checkPassword1();
+    let u = this.checkUsername();
+    let pwm = this.checkPasswordsMatching();
+    let r = this.checkRequired();
+    let cm = this.checkContactMethod();
+    let test = pw1 || u || pwm || !r || cm;
+    return test;
+  }
 
   register() {
     let registerLogin = this;
@@ -66,11 +233,13 @@ export class RegisterLoginPage {
     this.email = '';
     this.sms = '';
 
-    if (!this.terms) {
-      this.errors.push("You must accept Privacy and Terms to proceed.");
-      return;
-    }
+    this.usernameerrorvalue = '';
+    this.password1errorvalue = '';
+    this.password2errorvalue = '';
+    this.emailerrorvalue = '';
+    this.smserrorvalue = '';
 
+    // register object we're passing to the api
     let register = {
       username: this.username,
       password1: this.password1,
@@ -80,62 +249,86 @@ export class RegisterLoginPage {
       tandc: 1
     }
 
+    if (!this.terms) {
+      // this.errors.push("You must accept Privacy and Terms below to proceed.");
+      this.termserror = true;
+      this.content.scrollToBottom();
+      let scrollToBottomPromise = this.promiseToScroll();
+      let registerScope = this;
+      scrollToBottomPromise.then(function () {
+        registerScope.content.scrollToBottom();
+      });
+      this.checkAll();
+      return;
+    }
+    this.termserror = false;
+
     if (this.contactMethod.pcmethod === 'email') {
       this.email = this.contactMethod.pcvalue;
       register.email = this.email;
     }
-    else 
-    {
-      this.sms = '1'+this.contactMethod.mobileNumberAreaCode +
-      this.contactMethod.mobileNumberPrefix+
-      this.contactMethod.mobileNumberLineNumber;    
+    else {
+      this.sms = this.contactMethod.mobilenumber.getPN();
       register.phone = this.sms;
+      console.log('about to register with: ' + register.phone);
     }
-   
+
     if (!register.email) delete register.email;
     if (!register.phone) delete register.phone;
-    
+    if(!this.checkAll()) {
     this.userServices.register(register)
       .subscribe(
-      key => {
-        this.key = key;
-        let myProfile = {
-          'User': registerLogin.username
-        }
-        this.userServices.user.name = this.username;
-        
-        if (this.remember) {
-          this.storage.set('username', this.username);
-          this.storage.set('key', this.userServices.user.id);
-        }
-        else this.storage.set('username', '');
-       // registerLogin.nav.insert(registerLogin.nav.length(),HomePage);
-        registerLogin.nav.setPages([HomePage,RegisterIndividualProfilePage]);
+        key => {
+          this.key = key;
+          this.userServices.user.name = this.username;
 
+          if (this.remember) {
+            this.storage.set('username', this.username);
+            this.storage.set('key', this.userServices.user.id);
+          }
+          else this.storage.set('username', '');
 
-        /* this.userServices.createMyProfile(myProfile)
-           .subscribe(
-           key => {
-             registerLogin.key = key;
-             registerLogin.nav.setRoot(RegisterIndividualProfilePage);
-           },
-           err => {
-             console.log(err);
-             this.setError(err);
-           }),
-           val => this.val = val;
-           */
-      },
-      err => {
-        console.log(err);
-        this.setError(err);
-      });
+          if(this.navParams.get('fromPage')){
+            this.navCtrl.pop().then(() => {
+          // Trigger custom event and pass data to be send back
+          this.ev.publish('user-event-flow', this.navParams.get("event_id"));
+          //new
+         });
+       }else{
+           registerLogin.nav.setPages([{page: HomePage}, {page:RegisterIndividualProfilePage}]);
+       }
 
+        },
+        err => {
+          console.log(err);
+          this.setError(err);
+        });
+    }
   }
 
   back() {
     this.nav.popToRoot();
   }
+
+  openModal() {
+    let modal = this.modalCtrl.create(PrivacyTermsModal);
+    modal.onDidDismiss(data => {
+      console.log(data);
+      this.modalClicked = true;
+      this.terms = data.agree;
+      if (!this.terms) {
+        this.termserror = true;
+        let scrollToBottomPromise = this.promiseToScroll();
+        let registerScope = this;
+        scrollToBottomPromise.then(function () {
+          registerScope.content.scrollToBottom();
+        });
+      }
+      this.checkRequired();
+    });
+    modal.present();
+  }
+
 
   setError(error) {
     if (error.status === 400) {
@@ -145,43 +338,72 @@ export class RegisterLoginPage {
           let field = '';
           if (STRINGS[key]) field = STRINGS[key] + ': ';
           this.errors.push(field + error[key][val].toString());
-          if (key === 'username') this.usernameerror = true;
-          if (key === 'password1') this.password1error = true;
-          if (key === 'password2') this.password2error = true;
-          if (key === 'email') this.emailerror = true;
-          if (key === 'sms') this.smserror = true;
+          var message = error[key][val].toString();
+          if (key === 'username') {
+            this.usernameerror = true;
+            this.usernameerrorvalue = message;
+          }
+          if (key === 'password1') {
+            this.password1error = true;
+            this.password1errorvalue = message;
+          }
+          if (key === 'password2') {
+            this.password2error = true;
+            this.password2errorvalue = message;
+          }
+          if (key === 'email') {
+            this.emailerror = true;
+            this.emailerrorvalue = message;
+          }
+          if (key === 'phone') {
+            this.smserror = true;
+            this.smserrorvalue = message;
+          }
+          if (key === 'non_field_errors') {
+            this.nonFieldsError = true;
+            this.nonFieldsErrorValue = message;
+          }
         }
       }
+      // scroll to top of page
+      let promToScroll = this.promiseToScroll();
+      let registerScope = this;
+      promToScroll.then(() => {
+        registerScope.content.scrollToTop();
+      });
     }
     if (error.status === 500) {
       this.errors.push('Backend returned 500 error, talk to JOHN :) ');
     }
-
   }
+
   viewTerms() {
-    this.nav.push(TermsPage);
+    this.openModal();
   }
 
   showPassword() {
     if (this.showpassword === 'password') this.showpassword = 'text';
     else this.showpassword = 'password';
   }
-  presentUserPopover(ev) {
 
+  presentUserPopover(ev) {
     let popover = this.popoverCtrl.create(UseridPopover, {
     });
-
     popover.present({
       ev: ev
     });
   }
-  presentPasswordPopover(ev) {
 
+  presentPasswordPopover(ev) {
     let popover = this.popoverCtrl.create(PasswordPopover, {
     });
-
     popover.present({
       ev: ev
     });
   }
 }
+
+
+
+// WEBPACK FOOTER //
+// ./src/pages/register-login/register-login.ts
